Refactor sign-up into its own service
