+++ packages/twenty-server/src/core/auth/auth.module.ts
@@ -19,6 +19,7 @@ import { VerifyAuthController } from 'src/core/auth/controllers/verify-auth.cont
 import { TokenService } from 'src/core/auth/services/token.service';
 import { GoogleGmailService } from 'src/core/auth/services/google-gmail.service';
 import { UserWorkspaceModule } from 'src/core/user-workspace/user-workspace.module';
+import { SignUpService } from 'src/core/auth/services/sign-up.service';
 
 import { AuthResolver } from './auth.resolver';
 
@@ -54,6 +55,7 @@ const jwtModule = JwtModule.registerAsync({
     VerifyAuthController,
   ],
   providers: [
+    SignUpService,
     AuthService,
     TokenService,
     JwtAuthStrategy,
