+++ packages/twenty-server/src/core/auth/services/auth.service.ts
@@ -5,15 +5,10 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { HttpService } from '@nestjs/axios';
 
 import { Repository } from 'typeorm';
-import { v4 } from 'uuid';
 import { render } from '@react-email/components';
 import { PasswordUpdateNotifyEmail } from 'twenty-emails';
-import FileType from 'file-type';
-
-import { FileFolder } from 'src/core/file/interfaces/file-folder.interface';
 
 import { ChallengeInput } from 'src/core/auth/dto/challenge.input';
 import { assert } from 'src/utils/assert';
@@ -28,13 +23,10 @@ import { WorkspaceInviteHashValid } from 'src/core/auth/dto/workspace-invite-has
 import { User } from 'src/core/user/user.entity';
 import { Workspace } from 'src/core/workspace/workspace.entity';
 import { UserService } from 'src/core/user/services/user.service';
-import { WorkspaceManagerService } from 'src/workspace/workspace-manager/workspace-manager.service';
-import { FileUploadService } from 'src/core/file/services/file-upload.service';
 import { EnvironmentService } from 'src/integrations/environment/environment.service';
 import { EmailService } from 'src/integrations/email/email.service';
 import { UpdatePassword } from 'src/core/auth/dto/update-password.entity';
-import { getImageBufferFromUrl } from 'src/utils/image';
-import { UserWorkspaceService } from 'src/core/user-workspace/user-workspace.service';
+import { SignUpService } from 'src/core/auth/services/sign-up.service';
 
 import { TokenService } from './token.service';
 
@@ -49,14 +41,11 @@ export class AuthService {
   constructor(
     private readonly tokenService: TokenService,
     private readonly userService: UserService,
-    private readonly workspaceManagerService: WorkspaceManagerService,
-    private readonly fileUploadService: FileUploadService,
+    private readonly signUpService: SignUpService,
     @InjectRepository(Workspace, 'core')
     private readonly workspaceRepository: Repository<Workspace>,
     @InjectRepository(User, 'core')
     private readonly userRepository: Repository<User>,
-    private readonly userWorkspaceService: UserWorkspaceService,
-    private readonly httpService: HttpService,
     private readonly environmentService: EnvironmentService,
     private readonly emailService: EmailService,
   ) {}
@@ -94,117 +83,14 @@ export class AuthService {
     workspaceInviteHash?: string | null;
     picture?: string | null;
   }) {
-    if (!firstName) firstName = '';
-    if (!lastName) lastName = '';
-
-    const existingUser = await this.userRepository.findOne({
-      where: {
-        email: email,
-      },
-      relations: ['defaultWorkspace'],
+    return await this.signUpService.signUp({
+      email,
+      password,
+      firstName,
+      lastName,
+      workspaceInviteHash,
+      picture,
     });
-
-    if (existingUser && !workspaceInviteHash) {
-      assert(!existingUser, 'This user already exists', ForbiddenException);
-    }
-
-    if (password) {
-      const isPasswordValid = PASSWORD_REGEX.test(password);
-
-      assert(isPasswordValid, 'Password too weak', BadRequestException);
-    }
-
-    const passwordHash = password ? await hashPassword(password) : undefined;
-    let workspace: Workspace | null;
-
-    if (workspaceInviteHash) {
-      workspace = await this.workspaceRepository.findOneBy({
-        inviteHash: workspaceInviteHash,
-      });
-
-      assert(
-        workspace,
-        'This workspace inviteHash is invalid',
-        ForbiddenException,
-      );
-    } else {
-      assert(
-        !this.environmentService.isSignUpDisabled(),
-        'Sign up is disabled',
-        ForbiddenException,
-      );
-
-      const workspaceToCreate = this.workspaceRepository.create({
-        displayName: '',
-        domainName: '',
-        inviteHash: v4(),
-        subscriptionStatus: 'incomplete',
-      });
-
-      workspace = await this.workspaceRepository.save(workspaceToCreate);
-    }
-
-    let imagePath: string | undefined = undefined;
-
-    if (picture) {
-      const buffer = await getImageBufferFromUrl(
-        picture,
-        this.httpService.axiosRef,
-      );
-
-      const type = await FileType.fromBuffer(buffer);
-
-      const { paths } = await this.fileUploadService.uploadImage({
-        file: buffer,
-        filename: `${v4()}.${type?.ext}`,
-        mimeType: type?.mime,
-        fileFolder: FileFolder.ProfilePicture,
-      });
-
-      imagePath = paths[0];
-    }
-
-    if (existingUser && workspaceInviteHash) {
-      const userWorkspaceExists =
-        await this.userWorkspaceService.checkUserWorkspaceExists(
-          existingUser.id,
-          workspace.id,
-        );
-
-      if (!userWorkspaceExists) {
-        await this.userWorkspaceService.create(existingUser.id, workspace.id);
-
-        await this.userWorkspaceService.createWorkspaceMember(
-          workspace.id,
-          existingUser,
-        );
-      }
-
-      const updatedUser = await this.userRepository.save({
-        id: existingUser.id,
-        defaultWorkspace: workspace,
-        updatedAt: new Date().toISOString(),
-      });
-
-      return Object.assign(existingUser, updatedUser);
-    }
-
-    const userToCreate = this.userRepository.create({
-      email: email,
-      firstName: firstName,
-      lastName: lastName,
-      defaultAvatarUrl: imagePath,
-      canImpersonate: false,
-      passwordHash,
-      defaultWorkspace: workspace,
-    });
-
-    const user = await this.userRepository.save(userToCreate);
-
-    await this.userWorkspaceService.create(user.id, workspace.id);
-    await this.userWorkspaceService.createWorkspaceMember(workspace.id, user);
-
-    return user;
   }
 
   async verify(email: string): Promise<Verify> {
